댓글 제외 완성

+++ src/main/java/controller/commuController.java
@@ -11,6 +11,7 @@ import javax.servlet.annotation.WebServlet;
 import javax.servlet.http.HttpServlet;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
+import javax.servlet.http.HttpSession;
 
 import com.oreilly.servlet.MultipartRequest;
 
@@ -18,6 +19,7 @@ import domain.commu.dto.EditReqDto;
 import domain.commu.dto.ListReqDto;
 import domain.commu.dto.ViewReqDto;
 import domain.commu.dto.WriteReqDto;
+import domain.user.User;
 import service.commuService;
 
 
@@ -32,6 +34,7 @@ public class commuController extends HttpServlet {
     protected void process (HttpServletRequest req, HttpServletResponse res) throws ServletException, IOException {
     	String cmd = req.getParameter("cmd");
     	commuService cs = new commuService();
+    	HttpSession session = req.getSession();
     	//커뮤니티 글쓰기 페이지 요청
     	if (cmd.equals("write")) {
     		//filter로 페이지 close진행 예정으로 내부페이지 이동 가능한 getRequestDispatcher로 진행
@@ -56,22 +59,26 @@ public class commuController extends HttpServlet {
     		File nfile = new File(savePoint+File.separator+newFileName);
     		ofile.renameTo(nfile);
     		//parameter 값 받기
-    		String commuName = mr.getParameter("commuName");
+    		String title = mr.getParameter("title");
     		String scontent = mr.getParameter("scontent");
     		String lcontent = mr.getParameter("lcontent");
     		String id = mr.getParameter("userId");
+    		String address = mr.getParameter("address");
     		System.out.println(id);
     		WriteReqDto dto = new WriteReqDto();
-    		dto.setCommuName(commuName);
+    		dto.setTitle(title);
     		dto.setScontent(scontent);
     		dto.setLcontent(lcontent);
     		dto.setOfile(newFileName);
     		dto.setId(id);
+    		dto.setAddress(address);
     		
     		int result = cs.regiCommu(dto);
     		if (result==1) {
     			System.out.println("등록완료");
-    			req.getRequestDispatcher("commu?cmd=view&num="+dto.getNum()).forward(req, res);
+    			int rownum = cs.detailNum();
+    			req.getRequestDispatcher("commu?cmd=view&num="+rownum)
+    			.forward(req, res);
     		}else {
     			System.out.println("등록실패");
     		}
@@ -82,7 +89,7 @@ public class commuController extends HttpServlet {
     		int page = Integer.parseInt(req.getParameter("page"));
     		req.setAttribute("page", page);
     		List<ListReqDto> lists = cs.list(page);
-    		int lastPage = (cs.lastPage()-1)/5;
+    		int lastPage = (cs.lastPage()-1)/8;
     		req.setAttribute("lists", lists);
     		req.setAttribute("lastPage", lastPage);
     		req.getRequestDispatcher("community/community_List.jsp").forward(req, res);
@@ -91,8 +98,17 @@ public class commuController extends HttpServlet {
     	else if (cmd.equals("view")) {
     		int num = Integer.parseInt(req.getParameter("num"));
     		ViewReqDto dto = cs.detail(num);
-    		req.setAttribute("dto", dto);
-    		req.getRequestDispatcher("community/community_View.jsp").forward(req, res);
+    		//조회수 증가
+    		if(session.getAttribute("principal") != null) {
+    			User id = (User) session.getAttribute("principal");
+    			if(!id.getId().equals(dto.getId())) {
+        			System.out.println(id.getId() +"," + dto.getId());
+        			dto.setViews(cs.visitUpdate(num));
+        		}
+    		}
+    		req.setAttribute("dto",dto);
+    		req.getRequestDispatcher("community/community_View.jsp")
+    		.forward(req,res);
     	}
     	//수정하기 페이지 요청
     	else if(cmd.equals("edit")) {
@@ -109,29 +125,39 @@ public class commuController extends HttpServlet {
     		int maxSize = 5*1024*1024;
     		MultipartRequest mr = new MultipartRequest(req, savePoint, maxSize, "utf-8");
     		String change_img = mr.getParameter("change_img");
-    		File delete_img = new File(savePoint+File.separator+change_img);
-    		if (delete_img.exists()) {
-    			delete_img.delete();
-    		}
+    		String newFileName = null;
     		String ofile = mr.getFilesystemName("ofile");
-    		String exe = ofile.substring(ofile.lastIndexOf("."));
-    		String now = new SimpleDateFormat("yyyyMMdd_HHmmss").format(new Date());
-    		String newFileName = now+exe;
-    		File oldFile = new File(savePoint + File.separator + ofile);
-    		File newFile = new File(savePoint + File.separator + newFileName);
-    		oldFile.renameTo(newFile);
+    		if (ofile == null) {
+    			newFileName = change_img;
+    		}
+    		else {
+    			File delete_img = new File(savePoint + File.separator + change_img);
+        		if(delete_img.exists()) {
+        			delete_img.delete();
+        		}
+
+        		String exe = ofile.substring(ofile.lastIndexOf("."));
+        		String now = new SimpleDateFormat("yyyy-MM-dd_HHmmss").format(new Date());
+        		newFileName = now + exe;
+        		File oldFile = new File(savePoint + File.separator + ofile);
+        		System.out.println(oldFile);
+        		File newFile = new File(savePoint + File.separator + newFileName);
+        		System.out.println(newFile);
+        		oldFile.renameTo(newFile);
+    		}
+    		
     		
     		int num = Integer.parseInt(mr.getParameter("num"));
-    		String commuName = mr.getParameter("commuName");
+    		String title = mr.getParameter("title");
     		String scontent = mr.getParameter("scontent");
     		String lcontent = mr.getParameter("lcontent");
     		
     		ViewReqDto dto = new ViewReqDto();
     		dto.setNum(num);
-    		dto.setCommuName(commuName);
+    		dto.setTitle(title);
     		dto.setScontent(scontent);
     		dto.setLcontent(lcontent);
-    		dto.setOfile(ofile);
+    		dto.setOfile(newFileName);
     		
     		int result = cs.edit_process(dto);
     		if (result==1) {
@@ -140,8 +166,22 @@ public class commuController extends HttpServlet {
     			.forward(req, res);
     		}
     	}
+    	
+    	//게시글 삭제
+    	else if(cmd.equals("delete")) {
+    		System.out.println("커뮤니티 게시글 삭제");
+    		int num = Integer.parseInt(req.getParameter("num"));
+    		int result = cs.delete(num);
+    		if(result == 1) {
+    			System.out.println("커뮤니티 게시글 삭제 완료");
+    			req.getRequestDispatcher("commu?cmd=list&page=0")
+    			.forward(req, res);
+    			
+    		}
+    	}
     }
-	protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
+    
+    protected void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
 		process(request, response);
 	}
